renaming schemes to messages
