+++ tchannel-core/src/main/java/com/uber/tchannel/api/TChannel.java
@@ -35,9 +35,9 @@ import com.uber.tchannel.handlers.PingHandler;
 import com.uber.tchannel.handlers.RequestRouter;
 import com.uber.tchannel.handlers.ResponseRouter;
 import com.uber.tchannel.headers.ArgScheme;
-import com.uber.tchannel.schemes.JSONSerializer;
-import com.uber.tchannel.schemes.Serializer;
-import com.uber.tchannel.schemes.ThriftSerializer;
+import com.uber.tchannel.messages.JSONSerializer;
+import com.uber.tchannel.messages.Serializer;
+import com.uber.tchannel.messages.ThriftSerializer;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.channel.ChannelFuture;
