+++ tchannel-core/src/main/java/com/uber/tchannel/handlers/MessageDefragmenter.java
@@ -33,10 +33,10 @@ import com.uber.tchannel.frames.ErrorFrame;
 import com.uber.tchannel.frames.Frame;
 import com.uber.tchannel.headers.ArgScheme;
 import com.uber.tchannel.headers.TransportHeaders;
-import com.uber.tchannel.schemes.ErrorResponse;
-import com.uber.tchannel.schemes.Response;
-import com.uber.tchannel.schemes.Request;
-import com.uber.tchannel.schemes.TChannelMessage;
+import com.uber.tchannel.messages.ErrorResponse;
+import com.uber.tchannel.messages.Response;
+import com.uber.tchannel.messages.Request;
+import com.uber.tchannel.messages.TChannelMessage;
 import com.uber.tchannel.utils.TChannelUtilities;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
