+++ tchannel-core/src/main/java/com/uber/tchannel/handlers/MessageFragmenter.java
@@ -28,9 +28,9 @@ import com.uber.tchannel.fragmentation.FragmentationState;
 import com.uber.tchannel.codecs.TFrame;
 import com.uber.tchannel.frames.CallRequestFrame;
 import com.uber.tchannel.frames.CallResponseFrame;
-import com.uber.tchannel.schemes.Response;
-import com.uber.tchannel.schemes.RawMessage;
-import com.uber.tchannel.schemes.Request;
+import com.uber.tchannel.messages.Response;
+import com.uber.tchannel.messages.RawMessage;
+import com.uber.tchannel.messages.Request;
 import com.uber.tchannel.tracing.Trace;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
