+++ tchannel-core/src/main/java/com/uber/tchannel/handlers/RequestRouter.java
@@ -33,11 +33,11 @@ import com.uber.tchannel.api.handlers.RequestHandler;
 import com.uber.tchannel.errors.ErrorType;
 import com.uber.tchannel.headers.ArgScheme;
 import com.uber.tchannel.headers.TransportHeaders;
-import com.uber.tchannel.schemes.JSONSerializer;
-import com.uber.tchannel.schemes.Request;
-import com.uber.tchannel.schemes.ResponseMessage;
-import com.uber.tchannel.schemes.Serializer;
-import com.uber.tchannel.schemes.ThriftSerializer;
+import com.uber.tchannel.messages.JSONSerializer;
+import com.uber.tchannel.messages.Request;
+import com.uber.tchannel.messages.ResponseMessage;
+import com.uber.tchannel.messages.Serializer;
+import com.uber.tchannel.messages.ThriftSerializer;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 import io.netty.util.CharsetUtil;
