+++ tchannel-core/src/main/java/com/uber/tchannel/messages/RawMessage.java
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.uber.tchannel.schemes;
+package com.uber.tchannel.messages;
 
 import io.netty.buffer.ByteBuf;
 import java.util.Map;
