+++ tchannel-core/src/main/java/com/uber/tchannel/messages/Request.java
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.uber.tchannel.schemes;
+package com.uber.tchannel.messages;
 
 import com.uber.tchannel.frames.FrameType;
 import com.uber.tchannel.headers.ArgScheme;
