+++ tchannel-core/src/main/java/com/uber/tchannel/messages/ResponseMessage.java
@@ -20,7 +20,7 @@
  * THE SOFTWARE.
  */
 
-package com.uber.tchannel.schemes;
+package com.uber.tchannel.messages;
 
 import com.uber.tchannel.frames.FrameType;
 
@@ -29,8 +29,8 @@ import com.uber.tchannel.frames.FrameType;
  * <p>
  * All RPC frames over TChannel contain 3 opaque byte payloads, namely, arg{1,2,3}. TChannel makes no assumptions
  * about the contents of these frames. In order to make sense of these arg payloads, TChannel has the notion of
- * `arg schemes` which define standardized schemas and serialization formats over the raw arg{1,2,3} payloads. The
- * supported `arg schemes` are `thrift`, `json`, `http` and `sthrift`. These request / response frames will be built
+ * `arg messages` which define standardized schemas and serialization formats over the raw arg{1,2,3} payloads. The
+ * supported `arg messages` are `thrift`, `json`, `http` and `sthrift`. These request / response frames will be built
  * on top of {@link RawRequest} and {@link ResponseMessage} frames.
  * <p>
  * <h3>From the Docs</h3>
