+++ tchannel-core/src/test/java/com/uber/tchannel/handlers/FrameFragmenterTest.java
@@ -24,7 +24,7 @@ package com.uber.tchannel.handlers;
 
 import com.uber.tchannel.frames.CallFrame;
 import com.uber.tchannel.frames.CallRequestFrame;
-import com.uber.tchannel.schemes.RawRequest;
+import com.uber.tchannel.messages.RawRequest;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.embedded.EmbeddedChannel;
