+++ tchannel-core/src/test/java/com/uber/tchannel/handlers/TestMessageMultiplexer.java
@@ -27,9 +27,8 @@ import com.uber.tchannel.frames.CallRequestFrame;
 import com.uber.tchannel.frames.CallRequestContinueFrame;
 import com.uber.tchannel.headers.ArgScheme;
 import com.uber.tchannel.headers.TransportHeaders;
-import com.uber.tchannel.schemes.RawMessage;
-import com.uber.tchannel.schemes.TChannelMessage;
-import io.netty.buffer.ByteBufUtil;
+import com.uber.tchannel.messages.RawMessage;
+import com.uber.tchannel.messages.TChannelMessage;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.embedded.EmbeddedChannel;
 import io.netty.util.CharsetUtil;
