+++ tchannel-example/src/main/java/com/uber/tchannel/thrift/GetValueRequestHandler.java
@@ -24,8 +24,8 @@ package com.uber.tchannel.thrift;
 
 import java.util.Map;
 import com.uber.tchannel.api.handlers.ThriftRequestHandler;
-import com.uber.tchannel.schemes.ThriftRequest;
-import com.uber.tchannel.schemes.ThriftResponse;
+import com.uber.tchannel.messages.ThriftRequest;
+import com.uber.tchannel.messages.ThriftResponse;
 import com.uber.tchannel.thrift.generated.KeyValue;
 import com.uber.tchannel.thrift.generated.NotFoundError;
