+++ tchannel-example/src/main/java/com/uber/tchannel/thrift/SetValueRequestHandler.java
@@ -25,8 +25,8 @@ package com.uber.tchannel.thrift;
 import java.util.Map;
 
 import com.uber.tchannel.api.handlers.ThriftRequestHandler;
-import com.uber.tchannel.schemes.ThriftRequest;
-import com.uber.tchannel.schemes.ThriftResponse;
+import com.uber.tchannel.messages.ThriftRequest;
+import com.uber.tchannel.messages.ThriftResponse;
 import com.uber.tchannel.thrift.generated.KeyValue;
 
 public class SetValueRequestHandler extends ThriftRequestHandler<KeyValue.setValue_args, KeyValue.setValue_result> {
